feat: add solution for day 12 part 1

+++ src/test/java/com/adventofcode/year2024/day11/Day11RiddleSolverTest.java
@@ -1,7 +1,6 @@
 package com.adventofcode.year2024.day11;
 
 import com.adventofcode.common.AdventOfCodeFileReader;
-import com.adventofcode.year2024.day10.Day10RiddleSolver;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -18,7 +17,7 @@ public class Day11RiddleSolverTest {
 
   @Test
   void testSolveRiddleTwo() {
-    int result = riddleSolver.solveRiddleTwo(fileReader.readFile("src/test/resources/2024/11.txt"));
+    long result = riddleSolver.solveRiddleTwo(fileReader.readFile("src/test/resources/2024/11.txt"));
     assertEquals(81, result);
   }
 }
